Deleted login.jsp
It's not part of current sprint. Made it on Vacation because I was bored and wanted to do some .jsp work and practice that. We might use the file in the future though, and we might make a login feature eventually, so I'll move it to it's own feature branch.